Enable html5mode
* Change scripts/deploy.js to point to -pages domains so that we can
repurpose the original ones for the proxies

* Ensure that we turn this on at runtime _only_ if this is an angular app,
ie, don’t do this for Cordova.

This still does not work because ionic chokes at the presence of the
`<base>` tag. Standard way to address this is to add `requireBase: false`
when enabling html5Mode, but this breaks all the paths we have in the app