Implement getPrivateCrowdstartById, use in getCrowdstartById
If a crowdstarter route is not available in the usual dump from `/crowdstart/status`,
we assume it is a private one, and look up the route and bid info, and then transform
in the usual manner. This is implemented in `getPrivateCrowdstartById`.

That function returns a promise, the least invasive approach to integrating this into
`getCrowdstartById` (see the parent commit). Further, to prevent Angular from being
extremely confused when watching the retval of `getCrowdstartById`, we remember the
last looked up route id, and ensure we return the last promise created, if the route ids match

Given that private crowdstarts are only invoked if the user actually knows the URL, this should
not prove to have too much impact to regular usage

TODO: Figure out if there’s an even better way to wrap this up. Odds seem pretty low here.