RoutesService: Re-use inflight requests for route info (#710)
* RoutesService: Re-use inflight requests for route info

The frontend currently makes multiple requests for the same route info,
with `ignoreCache` specified as true. This is because the cache is
too crude in its current form, holding data indefinitely until the
next request.

This is not ideal as it makes triple or quadruple the number of
needed requests in one single load, due to the way the app is
currently built and structured, such that multiple points in the
app are effectively asking for the same thing. This in turn leads
to potential capacity issues in the backend.

To combat this, direct requests for route info to a common k-v map,
keying urls to the corresponding inflight request, making a new request
only if one is not currently inflight.

* Declare a new map, `inFlightRouteRequests`
* Modify `getRoute()` and `fetchRoutes()` such that if a request
  for a given url is already inflight, use that, otherwise, create
  a new one and register it against `inFlightRouteRequests`
* Take care to remove the request on completion