Consolidate lite route query to daily trip broker
Both the controller for details on lite routes and the daily trips
broker directive that the corresponding view embeds make their own
queries for lite route. Delegate that solely to the broker, using
an EventEmitter to pass any queried routes to the controller. This
then allows the controller to get the route info without making its
own queries